BUGFIX: Loading unloaded children does not keep linkableNodeTypes state
The configuration option allowedNodeTypes (internal name `linkableNodeTypes`) is not preserved after the initial tree was loaded. This option determines which node types are allowed to be linked (introduced in #11).

Likely a regression during the refactoring https://github.com/sitegeist/Sitegeist.Archaeopteryx/pull/63 - the PHP Api was already set but not used from the client.