TASK: #87 Migrate jest to node --test
Jest seems to be rotting a little and its also a little a thorn in the eyes as it still uses babel.
The build stack was refactored to use no longer babel and webpack but esbuild.

Node now also supports writing native test and for older more common node versions (e.g. 18) it is good practice to use tsx which puts esbuild infront of node before running the files. This is perfect in that the test setup would use the same tooling as the build pipeline. https://nodejs.org/en/learn/typescript/run#running-typescript-code-with-tsx

We should consider raising the node version at some point because Node 18 does not provide all the nice to have features:
- build in typescript support
- custom globbing (that we dont need to name our files .test.js)