BUGFIX: Re-fix node links with anchors #59
Reverts https://github.com/sitegeist/Sitegeist.Archaeopteryx/pull/60 partially

Node links can also contain an anchor. This is now parsed correctly in the InspectorEditor.
When the anchor option is set for the link the anchor will also be shown in the dialog and kept for succeeding changes.

```yaml
someLink:
  type: string
  ui:
    inspector:
      group: settings
      editor: Sitegeist.Archaeopteryx/Inspector/Editors/LinkEditor
      editorOptions:
        anchor: true
```

We no longer require the workaround to make the CustomLinkSpecification match the anchor links.